Extract wire testing to reusable trait
