+++ src/test/php/com/mongodb/unittest/ReplicaSetTest.class.php
@@ -7,123 +7,18 @@ use unittest\{Assert, Values, Test};
 use util\{Date, UUID};
 
 class ReplicaSetTest {
-  const PRIMARY    = 'shard2.test:27017';
-  const SECONDARY1 = 'shard1.test:27017';
-  const SECONDARY2 = 'shard0.test:27017';
-  const SESSION    = '6ce334fc-4ce5-4d0c-be6a-1e34feebbad4';
-
-  /**
-   * Connect to a given replica set definition
-   *
-   * @param  [:var[]] $definition
-   * @param  string $readPreference
-   * @return com.mongodb.io.Protocol
-   */
-  private function connect($definition, $readPreference= 'primary') {
-    $conn= [];
-    foreach ($definition as $node => $replies) {
-      $conn[]= new TestingConnection($node, $replies);
-    }
-
-    return (new Protocol($conn, ['params' => ['readPreference' => $readPreference]]))->connect();
-  }
-
-  /**
-   * Returns map of addresses to server kinds
-   * 
-   * @param  com.mongodb.io.Protocol $proto
-   * @return [:?string]
-   */
-  private function connected($proto) {
-    return array_map(
-      function($conn) { return $conn->connected() ? $conn->server['$kind'] : null; },
-      $proto->connections()
-    );
-  }
-
-  /**
-   * Creates a hello reply
-   *
-   * @param  string $node
-   * @return [:var]
-   */
-  private function hello($node) {
-    return [
-      'responseFlags'   => 8,
-      'cursorID'        => 0,
-      'startingFrom'    => 0,
-      'numberReturned'  => 1,
-      'documents'       => [[
-        'topologyVersion'              => ['processId' => new ObjectId('6235b5ddda38998abb76bed3'), new Int64(6)],
-        'hosts'                        => [self::PRIMARY, self::SECONDARY1, self::SECONDARY2],
-        'setName'                      => 'atlas-test-shard-0',
-        'setVersion'                   => 8,
-        'isWritablePrimary'            => self::PRIMARY === $node,
-        'secondary'                    => self::PRIMARY !== $node,
-        'primary'                      => self::PRIMARY,
-        'tags'                         => [
-          'provider'     => 'AWS',
-          'region'       => 'EU_CENTRAL_1',
-          'nodeType'     => 'ELECTABLE',
-          'workloadType' => 'OPERATIONAL'
-        ],
-        'me'                           => $node,
-        'electionId'                   => new ObjectId('7fffffff0000000000000136'),
-        'lastWrite'                    => [
-          'opTime'            => ['ts' => new Timestamp(1647687293, 41), 't' => new Int64(310)],
-          'lastWriteDate'     => new Date('2022-03-19 10:54:53+0000'),
-          'majorityOpTime'    => ['ts' => new Timestamp(1647687293, 41), 't' => new Int64(310)],
-          'majorityWriteDate' => new Date('2022-03-19 10:54:53+0000'),
-        ],
-        'maxBsonObjectSize'            => 16777216,
-        'maxMessageSizeBytes'          => 48000000,
-        'maxWriteBatchSize'            => 100000,
-        'localTime'                    => new Date('2022-03-19 10:54:53+0000'),
-        'logicalSessionTimeoutMinutes' => 30,
-        'connectionId'                 => 873,
-        'minWireVersion'               => 0,
-        'maxWireVersion'               => 13,
-        'readOnly'                     => self::PRIMARY !== $node,
-        'ok'                           => 1,
-      ]]
-    ];
-  }
-
-  /**
-   * Creates an OK reply
-   *
-   * @return [:var]
-   */
-  private function ok() {
-    return ['flags' => 0, 'body' => ['ok' => 1]];
-  }
-
-  /**
-   * Creates a count reply
-   *
-   * @param  int $n
-   * @return [:var]
-   */
-  private function count($n) {
-    return [
-      'flags' => 0,
-      'body'  => [
-        'cursor' => ['firstBatch' => [['n' => $n]], 'id' => new Int64(0), 'ns' => 'test.entries'],
-        'ok'     => 1,
-      ],
-    ];
-  }
+  use WireTesting;
 
   #[Test]
   public function connections_when_first_is_primary() {
     $fixture= $this->connect([
-      self::PRIMARY    => [$this->hello(self::PRIMARY)],
-      self::SECONDARY1 => [$this->hello(self::SECONDARY1)],
-      self::SECONDARY2 => [$this->hello(self::SECONDARY2)],
+      self::$PRIMARY    => [$this->hello(self::$PRIMARY)],
+      self::$SECONDARY1 => [$this->hello(self::$SECONDARY1)],
+      self::$SECONDARY2 => [$this->hello(self::$SECONDARY2)],
     ]);
 
     Assert::equals(
-      [self::PRIMARY => TestingConnection::RSPrimary, self::SECONDARY1 => null, self::SECONDARY2 => null],
+      [self::$PRIMARY => TestingConnection::RSPrimary, self::$SECONDARY1 => null, self::$SECONDARY2 => null],
       $this->connected($fixture)
     );
   }
@@ -131,13 +26,13 @@ class ReplicaSetTest {
   #[Test]
   public function connections_when_first_is_secondary() {
     $fixture= $this->connect([
-      self::SECONDARY1 => [$this->hello(self::SECONDARY1)],
-      self::SECONDARY2 => [$this->hello(self::SECONDARY2)],
-      self::PRIMARY    => [$this->hello(self::PRIMARY)],
+      self::$SECONDARY1 => [$this->hello(self::$SECONDARY1)],
+      self::$SECONDARY2 => [$this->hello(self::$SECONDARY2)],
+      self::$PRIMARY    => [$this->hello(self::$PRIMARY)],
     ]);
 
     Assert::equals(
-      [self::SECONDARY1 => TestingConnection::RSSecondary, self::SECONDARY2 => null, self::PRIMARY => null],
+      [self::$SECONDARY1 => TestingConnection::RSSecondary, self::$SECONDARY2 => null, self::$PRIMARY => null],
       $this->connected($fixture)
     );
   }
@@ -145,41 +40,41 @@ class ReplicaSetTest {
   #[Test]
   public function connects_to_primary_when_no_secondary_available() {
     $replicaSet= [
-      self::PRIMARY    => [$this->hello(self::PRIMARY), $this->ok()],
-      self::SECONDARY1 => [],
-      self::SECONDARY2 => [],
+      self::$PRIMARY    => [$this->hello(self::$PRIMARY), $this->ok()],
+      self::$SECONDARY1 => [],
+      self::$SECONDARY2 => [],
     ];
     $fixture= $this->connect($replicaSet, 'secondaryPreferred');
     $fixture->read(null, [/* anything */]);
 
     Assert::equals(
-      [self::PRIMARY => TestingConnection::RSPrimary, self::SECONDARY1 => null, self::SECONDARY2 => null],
+      [self::$PRIMARY => TestingConnection::RSPrimary, self::$SECONDARY1 => null, self::$SECONDARY2 => null],
       $this->connected($fixture)
     );
   }
 
   #[Test, Expect(class: NoSuitableCandidate::class, withMessage: '/No suitable candidate eligible for initial connect/')]
   public function throws_exception_if_none_of_the_nodes_are_reachable() {
-    $this->connect([self::PRIMARY => null, self::SECONDARY1 => null, self::SECONDARY2 => null]);
+    $this->connect([self::$PRIMARY => null, self::$SECONDARY1 => null, self::$SECONDARY2 => null]);
   }
 
   #[Test, Expect(class: NoSuitableCandidate::class, withMessage: '/No suitable candidate eligible for initial connect/')]
   public function throws_exception_if_none_of_the_nodes_respond() {
-    $this->connect([self::PRIMARY => [], self::SECONDARY1 => [], self::SECONDARY2 => []]);
+    $this->connect([self::$PRIMARY => [], self::$SECONDARY1 => [], self::$SECONDARY2 => []]);
   }
 
   #[Test, Values(['secondary', 'secondaryPreferred'])]
   public function reads_from_first_secondary_with($readPreference) {
     $replicaSet= [
-      self::PRIMARY    => [$this->hello(self::PRIMARY)],
-      self::SECONDARY1 => [$this->hello(self::SECONDARY1), $this->ok()],
-      self::SECONDARY2 => [$this->hello(self::SECONDARY2)],
+      self::$PRIMARY    => [$this->hello(self::$PRIMARY)],
+      self::$SECONDARY1 => [$this->hello(self::$SECONDARY1), $this->ok()],
+      self::$SECONDARY2 => [$this->hello(self::$SECONDARY2)],
     ];
     $fixture= $this->connect($replicaSet, $readPreference);
     $fixture->read(null, [/* anything */]);
 
     Assert::equals(
-      [self::PRIMARY => TestingConnection::RSPrimary, self::SECONDARY1 => TestingConnection::RSSecondary, self::SECONDARY2 => null],
+      [self::$PRIMARY => TestingConnection::RSPrimary, self::$SECONDARY1 => TestingConnection::RSSecondary, self::$SECONDARY2 => null],
       $this->connected($fixture)
     );
   }
@@ -187,15 +82,15 @@ class ReplicaSetTest {
   #[Test, Values(['secondary', 'secondaryPreferred'])]
   public function reads_from_first_available_secondary_with($readPreference) {
     $replicaSet= [
-      self::PRIMARY    => [$this->hello(self::PRIMARY)],
-      self::SECONDARY1 => [],
-      self::SECONDARY2 => [$this->hello(self::SECONDARY2), $this->ok()],
+      self::$PRIMARY    => [$this->hello(self::$PRIMARY)],
+      self::$SECONDARY1 => [],
+      self::$SECONDARY2 => [$this->hello(self::$SECONDARY2), $this->ok()],
     ];
     $fixture= $this->connect($replicaSet, $readPreference);
     $fixture->read(null, [/* anything */]);
 
     Assert::equals(
-      [self::PRIMARY => TestingConnection::RSPrimary, self::SECONDARY1 => null, self::SECONDARY2 => TestingConnection::RSSecondary],
+      [self::$PRIMARY => TestingConnection::RSPrimary, self::$SECONDARY1 => null, self::$SECONDARY2 => TestingConnection::RSSecondary],
       $this->connected($fixture)
     );
   }
@@ -203,15 +98,15 @@ class ReplicaSetTest {
   #[Test, Values(['secondary', 'secondaryPreferred'])]
   public function reads_from_another_secondary($readPreference) {
     $replicaSet= [
-      self::PRIMARY    => [$this->hello(self::PRIMARY)],
-      self::SECONDARY1 => [$this->hello(self::SECONDARY1)],
-      self::SECONDARY2 => [$this->hello(self::SECONDARY2), $this->ok()],
+      self::$PRIMARY    => [$this->hello(self::$PRIMARY)],
+      self::$SECONDARY1 => [$this->hello(self::$SECONDARY1)],
+      self::$SECONDARY2 => [$this->hello(self::$SECONDARY2), $this->ok()],
     ];
     $fixture= $this->connect($replicaSet, $readPreference);
     $fixture->read(null, [/* anything */]);
 
     Assert::equals(
-      [self::PRIMARY => TestingConnection::RSPrimary, self::SECONDARY1 => null, self::SECONDARY2 => TestingConnection::RSSecondary],
+      [self::$PRIMARY => TestingConnection::RSPrimary, self::$SECONDARY1 => null, self::$SECONDARY2 => TestingConnection::RSSecondary],
       $this->connected($fixture)
     );
   }
@@ -219,9 +114,9 @@ class ReplicaSetTest {
   #[Test, Values(map: ['primary' => 45, 'primaryPreferred' => 45, 'secondary' => 44, 'secondaryPreferred' => 44, 'nearest' => 45])]
   public function reading_with($readPreference, $result) {
     $replicaSet= [
-      self::PRIMARY     => [$this->hello(self::PRIMARY), $this->count(45)],
-      self::SECONDARY1  => [$this->hello(self::SECONDARY1), $this->count(44)],
-      self::SECONDARY2  => [$this->hello(self::SECONDARY1), $this->count(44)],
+      self::$PRIMARY     => [$this->hello(self::$PRIMARY), $this->count(45)],
+      self::$SECONDARY1  => [$this->hello(self::$SECONDARY1), $this->count(44)],
+      self::$SECONDARY2  => [$this->hello(self::$SECONDARY1), $this->count(44)],
     ];
     $response= $this->connect($replicaSet, $readPreference)->read(null, [
       'aggregate' => 'test.entries',
@@ -236,7 +131,7 @@ class ReplicaSetTest {
   #[Test]
   public function writing() {
     $replicaSet= [
-      self::PRIMARY     => [$this->hello(self::PRIMARY), [
+      self::$PRIMARY     => [$this->hello(self::$PRIMARY), [
         'flags' => 0,
         'body'  => [
           'n'          => 1,
@@ -245,7 +140,7 @@ class ReplicaSetTest {
           'ok'         => 1
         ]
       ]],
-      self::SECONDARY1  => [$this->hello(self::SECONDARY1), [
+      self::$SECONDARY1  => [$this->hello(self::$SECONDARY1), [
         'flags' => 0,
         'body'  => [
           'topologyVersion' => ['processId' => new ObjectId('6235b49ff525c94b4ecdd835'), 'counter' => new Int64(4)],
@@ -255,7 +150,7 @@ class ReplicaSetTest {
           'codeName'        => 'NotWritablePrimary',
         ]
       ]],
-      self::SECONDARY2  => [],
+      self::$SECONDARY2  => [],
     ];
     $response= $this->connect($replicaSet)->write(null, [
       'delete'    => 'test',
@@ -270,9 +165,9 @@ class ReplicaSetTest {
   #[Test, Expect(class: NoSuitableCandidate::class, withMessage: '/No suitable candidate eligible for reading with secondary/')]
   public function read_throws_if_no_secondaries_are_available() {
     $replicaSet= [
-      self::PRIMARY    => [$this->hello(self::PRIMARY)],
-      self::SECONDARY1 => [$this->hello(self::SECONDARY1)],
-      self::SECONDARY2 => [$this->hello(self::SECONDARY2)],
+      self::$PRIMARY    => [$this->hello(self::$PRIMARY)],
+      self::$SECONDARY1 => [$this->hello(self::$SECONDARY1)],
+      self::$SECONDARY2 => [$this->hello(self::$SECONDARY2)],
     ];
     $fixture= $this->connect($replicaSet, 'secondary');
     $fixture->read(null, [/* anything */]);
@@ -281,9 +176,9 @@ class ReplicaSetTest {
   #[Test, Expect(class: NoSuitableCandidate::class, withMessage: '/No suitable candidate eligible for writing/')]
   public function write_throws_if_no_primary_is_available() {
     $replicaSet= [
-      self::PRIMARY    => [$this->hello(self::PRIMARY)],
-      self::SECONDARY1 => [$this->hello(self::SECONDARY1)],
-      self::SECONDARY2 => [$this->hello(self::SECONDARY2)],
+      self::$PRIMARY    => [$this->hello(self::$PRIMARY)],
+      self::$SECONDARY1 => [$this->hello(self::$SECONDARY1)],
+      self::$SECONDARY2 => [$this->hello(self::$SECONDARY2)],
     ];
     $fixture= $this->connect($replicaSet, 'primary');
     $fixture->write(null, [/* anything */]);
@@ -292,9 +187,9 @@ class ReplicaSetTest {
   #[Test]
   public function reconnects_when_checking_for_socket_with_ping_fails() {
     $replicaSet= [
-      self::PRIMARY    => [$this->hello(self::PRIMARY)],
-      self::SECONDARY1 => [$this->hello(self::SECONDARY1), $this->count(44), null, $this->hello(self::SECONDARY1), $this->count(45)],
-      self::SECONDARY2 => [],
+      self::$PRIMARY    => [$this->hello(self::$PRIMARY)],
+      self::$SECONDARY1 => [$this->hello(self::$SECONDARY1), $this->count(44), null, $this->hello(self::$SECONDARY1), $this->count(45)],
+      self::$SECONDARY2 => [],
     ];
     $fixture= $this->connect($replicaSet, 'secondary');
     $fixture->socketCheckInterval= 0;
@@ -322,7 +217,7 @@ class ReplicaSetTest {
   #[Test]
   public function does_not_disconnect_for_operation_errors() {
     $replicaSet= [
-      self::PRIMARY    => [$this->hello(self::PRIMARY), [
+      self::$PRIMARY    => [$this->hello(self::$PRIMARY), [
         'flags' => 0,
         'body'  => [
           'topologyVersion' => ['processId' => new ObjectId('6235b49ff525c94b4ecdd835'), 'counter' => new Int64(4)],
@@ -332,8 +227,8 @@ class ReplicaSetTest {
           'codeName'        => 'AtlasError',
         ]
       ]],
-      self::SECONDARY1 => [$this->hello(self::SECONDARY1)],
-      self::SECONDARY2 => [$this->hello(self::SECONDARY2)],
+      self::$SECONDARY1 => [$this->hello(self::$SECONDARY1)],
+      self::$SECONDARY2 => [$this->hello(self::$SECONDARY2)],
     ];
     $fixture= $this->connect($replicaSet, 'primary');
 
@@ -341,55 +236,8 @@ class ReplicaSetTest {
       $fixture->read(null, [/* anything */]);
     });
     Assert::equals(
-      [self::PRIMARY => TestingConnection::RSPrimary, self::SECONDARY1 => null, self::SECONDARY2 => null],
+      [self::$PRIMARY => TestingConnection::RSPrimary, self::$SECONDARY1 => null, self::$SECONDARY2 => null],
       $this->connected($fixture)
     );
-  }
-
-  #[Test]
-  public function session_id_is_sent_along() {
-    $replicaSet= [
-      self::PRIMARY    => [$this->hello(self::PRIMARY), $this->count(45), $this->ok()],
-      self::SECONDARY1 => [],
-      self::SECONDARY2 => [],
-    ];
-    $fixture= $this->connect($replicaSet, 'primary');
-    $count= ['count' => 'entries', '$db' => 'test'];
-
-    $id= new UUID(self::SESSION);
-    $session= new Session($fixture, $id);
-    $fixture->read($session, $count);
-    $session->close();
-
-    $conn= $fixture->connections()[self::PRIMARY];
-    $context= ['lsid' => ['id' => $id], '$readPreference' => ['mode' => 'primary']];
-
-    Assert::equals($count + $context, $conn->command(-2));
-    Assert::equals(['endSessions' => [['id' => $id]], '$db' => 'admin'] + $context, $conn->command(-1));
-  }
-
-  #[Test]
-  public function transaction_context_is_sent_along() {
-    $replicaSet= [
-      self::PRIMARY    => [$this->hello(self::PRIMARY), $this->ok(), $this->ok(), $this->ok()],
-      self::SECONDARY1 => [],
-      self::SECONDARY2 => [],
-    ];
-    $fixture= $this->connect($replicaSet, 'primary');
-    $update= ['update' => 'entries', 'updates' => [], '$db' => 'test'];
-
-    $id= new UUID(self::SESSION);
-    $transaction= (new Session($fixture, $id))->transaction();
-    $fixture->write($transaction, $update);
-    $fixture->write($transaction, $update);
-    $transaction->commit();
-
-    $conn= $fixture->connections()[self::PRIMARY];
-    $context= ['lsid' => ['id' => $id], '$readPreference' => ['mode' => 'primary']];
-    $txn= ['txnNumber' => new Int64(1), 'autocommit' => false];
-
-    Assert::equals($txn + ['startTransaction' => true] + $update + $context, $conn->command(-3));
-    Assert::equals($txn + $update + $context, $conn->command(-2));
-    Assert::equals($txn + ['commitTransaction' => 1, '$db' => 'admin'] + $context, $conn->command(-1));
   }
 }
